backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
python-dotenv==1.0.1
pydantic[email]==2.10.3
gspread==6.1.4
google-auth==2.37.0
aiosmtplib==3.0.2
